version 1.1
Added PAFit_oneshot function and fixing dois.

+++ R/from_igraph.R
@@ -2,7 +2,7 @@ from_igraph <- function(net) {
   if (!is(net,"igraph")) {
     stop("net must be an object of class igraph.")  
   }
-  graph <- matrix(as.integer(as.matrix(as_data_frame(net))), ncol = 3)
+  graph <- matrix(as.numeric(as.matrix(as_data_frame(net))), ncol = 3)
   if (is_directed(net))
     type <- "directed"
   else type <- "undirected"
